crackthedeck-backend/crackthedeck-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
reportlab>=4.1.0
pypdf>=4.0.0
pdf2image>=1.17.0
Pillow>=10.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0
feedparser>=6.0.0
